Move helpers from lib into our own utils
